backend/requirements.txt
django>=5.0,<6.0
djangorestframework>=3.14,<4.0
django-filter>=24.0
django-cors-headers>=4.0
